feat: 命令存储ES部分和日志存储ES部分代码优化
